+++ src/pages/workflows/model/workflow-list.ts
@@ -1,9 +1,10 @@
 import {
+  type NodeDefinitionResponse,
+  type WorkflowResponse,
   getDateTimestamp,
   getRelativeTimeLabel,
   getServiceBadgeKeyFromService,
 } from "@/shared";
-import type { NodeDefinitionResponse, WorkflowResponse } from "@/shared/api";
 
 import { type ServiceBadgeKey, type WorkflowFilterKey } from "./types";
