+++ src/shared/model/index.ts
@@ -1,5 +1,5 @@
-export * from "./useExecutionQueries";
-export * from "./useOAuthTokenQueries";
-export * from "./useTemplateQueries";
-export * from "./useWorkflowQueries";
+export * from "./execution";
+export * from "./oauth";
+export * from "./template";
+export * from "./workflow";
 export * from "./workflowStore";
