+++ src/shared/model/oauth/useConnectOAuthTokenMutation.ts
@@ -1,0 +1,8 @@
+import { useMutation } from "@tanstack/react-query";
+
+import { oauthApi } from "../../api";
+
+export const useConnectOAuthTokenMutation = () =>
+  useMutation({
+    mutationFn: (service: string) => oauthApi.connect(service),
+  });
