+++ src/shared/model/template/index.ts
@@ -1,0 +1,4 @@
+export * from "./useCreateTemplateMutation";
+export * from "./useInstantiateTemplateMutation";
+export * from "./useTemplateListQuery";
+export * from "./useTemplateQuery";
