--- src/shared/model/useOAuthTokenQueries.ts
@@ -1,27 +1,0 @@
-import { useMutation, useQuery } from "@tanstack/react-query";
-
-import { oauthApi } from "../api";
-import { oauthKeys } from "../constants";
-import { queryClient } from "../libs";
-
-export const useOAuthTokensQuery = () =>
-  useQuery({
-    queryKey: oauthKeys.tokens(),
-    queryFn: () => oauthApi.getTokens(),
-    throwOnError: false,
-  });
-
-export const useConnectOAuthTokenMutation = () =>
-  useMutation({
-    mutationFn: (service: string) => oauthApi.connect(service),
-  });
-
-export const useDisconnectOAuthTokenMutation = () =>
-  useMutation({
-    mutationFn: (service: string) => oauthApi.disconnect(service),
-    onSuccess: async () => {
-      await queryClient.invalidateQueries({
-        queryKey: oauthKeys.tokens(),
-      });
-    },
-  });
