+++ src/shared/model/workflow/useCreateWorkflowMutation.ts
@@ -1,0 +1,11 @@
+import { useMutation } from "@tanstack/react-query";
+
+import { workflowApi } from "../../api";
+
+import { syncWorkflowCache } from "./workflow-cache-utils";
+
+export const useCreateWorkflowMutation = () =>
+  useMutation({
+    mutationFn: workflowApi.create,
+    onSuccess: syncWorkflowCache,
+  });
