+++ src/shared/model/workflow/useDeleteWorkflowNodeMutation.ts
@@ -1,0 +1,17 @@
+import { useMutation } from "@tanstack/react-query";
+
+import { workflowApi } from "../../api";
+
+import { syncWorkflowCache } from "./workflow-cache-utils";
+
+export const useDeleteWorkflowNodeMutation = () =>
+  useMutation({
+    mutationFn: ({
+      workflowId,
+      nodeId,
+    }: {
+      workflowId: string;
+      nodeId: string;
+    }) => workflowApi.deleteNode(workflowId, nodeId),
+    onSuccess: syncWorkflowCache,
+  });
