+++ src/shared/model/workflow/useGenerateWorkflowMutation.ts
@@ -1,0 +1,11 @@
+import { useMutation } from "@tanstack/react-query";
+
+import { workflowApi } from "../../api";
+
+import { syncWorkflowCache } from "./workflow-cache-utils";
+
+export const useGenerateWorkflowMutation = () =>
+  useMutation({
+    mutationFn: workflowApi.generate,
+    onSuccess: syncWorkflowCache,
+  });
