+++ src/shared/types/api-models.type.ts
@@ -1,0 +1,10 @@
+export type {
+  ChoiceBranchConfig,
+  ChoiceFollowUp,
+  ChoiceOption,
+  ChoiceResponse,
+  NodeAddRequest,
+  NodeDefinitionResponse,
+  NodeSelectionResult,
+  WorkflowResponse,
+} from "../api/workflow";
