+++ src/shared/types/index.ts
@@ -1,1 +1,2 @@
 export * from "./api.type";
+export * from "./api-models.type";
